+++ src/main/java/one/pkg/modpublish/api/API.java
@@ -18,52 +18,45 @@
 package one.pkg.modpublish.api;
 
 import com.intellij.openapi.project.Project;
-import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.Response;
 import one.pkg.modpublish.data.internel.ModInfo;
 import one.pkg.modpublish.data.internel.PublishData;
 import one.pkg.modpublish.data.result.PublishResult;
-import one.pkg.modpublish.util.proxy.MProxy;
 import one.pkg.modpublish.util.resources.Lang;
 
 import java.util.Optional;
-import java.util.concurrent.TimeUnit;
 
-public interface API {
-    OkHttpClient client = new OkHttpClient.Builder().proxy(MProxy.getProxy())
-            .connectTimeout(20, TimeUnit.SECONDS)
-            .readTimeout(20, TimeUnit.SECONDS)
-            .writeTimeout(20, TimeUnit.SECONDS)
-            //.hostnameVerifier(SSLSocketClient.getHostnameVerifier())
-            //.sslSocketFactory(SSLSocketClient.getSSLSocketFactory(), SSLSocketClient.getX509TrustManager())
-            .build();
+public abstract class API {
 
-    default Request.Builder getBaseRequestBuilder() {
+    public API() {
+    }
+
+    Request.Builder getBaseRequestBuilder() {
         return new Request.Builder().header("User-Agent", "modpublish/v1 (github.com/404Setup/ModPublish)");
     }
 
-    String getID();
+    public abstract String getID();
 
-    void updateABServer();
+    public abstract void updateABServer();
 
-    boolean getABServer();
+    public abstract boolean getABServer();
 
-    default Request.Builder getJsonRequest(Request.Builder builder) {
+    Request.Builder getJsonRequest(Request.Builder builder) {
         return builder.header("Accept", "application/json");
     }
 
-    default Request.Builder getFormRequest(Request.Builder builder) {
+    Request.Builder getFormRequest(Request.Builder builder) {
         return builder.header("Content-Type", "multipart/form-data");
     }
 
-    default Optional<String> getContentType(Response response) {
+    Optional<String> getContentType(Response response) {
         return Optional.ofNullable(response.header("Content-Type"));
     }
 
-    String createJsonBody(PublishData data, Project project);
+    abstract String createJsonBody(PublishData data, Project project);
 
-    default Optional<String> getStatus(Response response) {
+    Optional<String> getStatus(Response response) {
         if (response.code() == 403)
             return Optional.of(Lang.get("api.common.err.403"));
         if (response.code() == 404)
@@ -84,7 +77,7 @@ public interface API {
         return Optional.empty();
     }
 
-    PublishResult createVersion(PublishData data, Project project);
+    public abstract PublishResult createVersion(PublishData data, Project project);
 
-    ModInfo getModInfo(String modid, Project project);
+    public abstract ModInfo getModInfo(String modid, Project project);
 }